--- src/app/product/product-edit/product-edit.component.html
@@ -1,84 +1,0 @@
-<div class="container">
-  <h1>{{title}}</h1>
-   <p style="font-weight: lighter;color: gray;font-size: 14px;" class="card-text">to reserve, please fill this
-     informations</p>
-     
-
-<mat-horizontal-stepper labelPosition="bottom" #stepper>
-    <mat-step [stepControl]="HospitalForm"> 
-
-    <form [formGroup]="HospitalForm">
-      <div class="form-row">
-        <input type="hidden" formControlName="id">
-        <div class="form-group col-md-6">
-          <input style="text-align: center;font-size: 14px;" type="text" class="form-control" formControlName="countryHealthId"  
-            placeholder="countryHealthId">
-        </div>
-        <div class="form-group col-md-6">
-          <input style="text-align: center;font-size: 14px;" type="text" class="form-control" formControlName="name"  
-            placeholder="name">
-        </div>  
-
-      <div class="form-group col-md-6">
-          <input style="text-align: center;font-size: 14px;" type="text" class="form-control" formControlName="remark"  
-            placeholder="REMARK">
-        </div> 
-    
-
-      <div class="form-group col-md-6">
-          <input style="text-align: center;font-size: 14px;" type="text" class="form-control" formControlName="history"  
-            placeholder="history">
-        </div>  
-
-      <div class="form-group col-md-6">
-          <input style="text-align: center;font-size: 14px;" type="text" class="form-control" formControlName="categoryName"  
-            placeholder="categoryName">
-        </div>  
-     
-       <div class="input-group mb-3">
-              <select formControlName="hospitalCategoryId" placeholder="City" class="custom-select" id="inputGroupSelect02">
-                <option *ngFor="let item of listhopitalCatValues" selected value="{{item.id}}">{{item.remark}}</option>
-              </select>
-        </div> 
-
-    </div>
-    </form>
-
-
-     <br> 
-
-   <button type="submit" class="btn btn-success btn-block" [disabled]='!(this.HospitalForm.valid && this.contactForm.valid)' (click)="reserve()">Reserve</button><br>
-
-    <div>
-          <button mat-button matStepperNext disabled="true" >Next</button>
-      </div>
-    </mat-step>
-
-    <mat-step [stepControl]="updatePuctureImage" optional > 
-      <form [formGroup]="updatePuctureImage">
-        <ng-template matStepLabel  >Fill out your images</ng-template>
-     
-        <div class="form-group col-md-6">  
-          <label for="exampleFormControlFile1">Example file input</label>
-          <input type="file" name="CovePathForm" (change)="onFileSelectCover($event)"  class="form-control-file" id="exampleFormControlFile1"> 
-        </div> 
-
-        
-        <div class="form-group col-md-6">  
-          <label for="exampleFormControlFile1">Example file input</label>
-          <input type="file" name="PictureProfilePathForm" (change)="onFileSelect($event)"  class="form-control-file" id="exampleFormControlFile1"> 
-        </div> 
-
-        
-       <button type="submit" class="btn btn-success btn-block" (click)="updateImages()">UpdateImages</button><br>
-
-        <div>
-          <button mat-button matStepperPrevious>Back</button>
-          <button mat-button matStepperNext>Next</button>
-        </div>
-      </form>
-    </mat-step> 
-
-  </mat-horizontal-stepper>
-    
- </div>
