--- src/app/product/product.model.ts
@@ -1,16 +1,0 @@
-import { Contact } from '../contacts/contact.model';
-import { GeneratedId } from '../generatedId';
-
-export class Product extends GeneratedId{
-    countryHealthId: string;
-    name: string;
-    remark: string;
-    history: string;
-    hospitalCategoryId: string;
-    categoryName: string;
-    covePath :  string;
-    pictureProfilePath : string;
-    contactModel:Contact;  
-}
-
-
