--- src/app/product/product/product.component.css
@@ -1,5 +1,0 @@
- 
-table {
-    width: 100%;
-    position: relative;    
-  }
