--- src/app/product/product/product.component.html
@@ -1,57 +1,0 @@
- 
-
-<mat-form-field class="example-full-width" style="float: right;">
-  <input (keyup)="applyFilter($event.target.value)" matInput placeholder="Filter">
-</mat-form-field>
-
-<div style="float: left;">
-<app-menu 
-[IsRowSelected]="IsRowSelected" 
-[addAction]="add" 
-[editAction]="edit" 
-[deleteAction]="delete" 
-[IsMultple]="IsMultple"
-[isNew]="isNew" 
-></app-menu>
-</div>
-<table mat-table [dataSource]="dataSource" matSort class="mat-elevation-z8">
-  <ng-container matColumnDef="select">
-    <th mat-header-cell *matHeaderCellDef>
-      <mat-checkbox (change)="$event ? masterToggle() : null; onrowselect()" [checked]="selection.hasValue() && isAllSelected()"
-        [indeterminate]="selection.hasValue() && !isAllSelected()" [aria-label]="checkboxLabel()"    >
-      </mat-checkbox>
-    </th>
-    <td mat-cell *matCellDef="let row">
-      <mat-checkbox (click)="$event.stopPropagation()" (change)="$event ? selection.toggle(row) : null; onrowselect()"
-        [checked]="selection.isSelected(row)" [aria-label]="checkboxLabel(row)" (click)="onrowselect()" >
-      </mat-checkbox>
-    </td>
-  </ng-container>
-
-  <ng-container matColumnDef="countryHealthId">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header> countryHealthId </th>
-    <td mat-cell *matCellDef="let element"> {{element.countryHealthId}} </td>
-  </ng-container>
-
-  <ng-container matColumnDef="name">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header> name </th>
-    <td mat-cell *matCellDef="let element"> {{element.name}} </td>
-  </ng-container>
-
-  <ng-container matColumnDef="remark">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header> remark </th>
-    <td mat-cell *matCellDef="let element"> {{element.remark}} </td>
-  </ng-container>
-
-  <ng-container matColumnDef="history">
-    <th mat-header-cell *matHeaderCellDef mat-sort-header> history </th>
-    <td mat-cell *matCellDef="let element"> {{element.history}} </td>
-  </ng-container>
-
-  <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-  <tr mat-row *matRowDef="let row; columns: displayedColumns;" (click)="selection.toggle(row);"
-    (click)="onrowselect(row)" ></tr>
-
-
-</table>
-<mat-paginator [pageSizeOptions]="[5, 10, 20]" showFirstLastButtons></mat-paginator>
