--- src/app/product/store/app-state.ts
@@ -1,1 +1,0 @@
-export interface AppState{}
