+++ src/app/products/product/product.component.html
@@ -1,1 +1,78 @@
-<p>product works!</p>
+ 
+
+<mat-form-field class="example-full-width" style="float: right;">
+  <input (keyup)="applyFilter($event.target.value)" matInput placeholder="Filter">
+</mat-form-field>
+
+<div style="float: left;">
+<app-menu 
+[IsRowSelected]="IsRowSelected" 
+[addAction]="add" 
+[editAction]="edit" 
+[deleteAction]="delete" 
+[IsMultple]="IsMultple"
+[isNew]="isNew" 
+></app-menu>
+</div>
+
+
+<table mat-table [dataSource]="dataSource" matSort class="mat-elevation-z8">
+  <ng-container matColumnDef="select">
+    <th mat-header-cell *matHeaderCellDef>
+      <mat-checkbox (change)="$event ? masterToggle() : null; onrowselect()" [checked]="selection.hasValue() && isAllSelected()"
+        [indeterminate]="selection.hasValue() && !isAllSelected()" [aria-label]="checkboxLabel()"    >
+      </mat-checkbox>
+    </th>
+    <td mat-cell *matCellDef="let row">
+      <mat-checkbox (click)="$event.stopPropagation()" (change)="$event ? selection.toggle(row) : null; onrowselect()"
+        [checked]="selection.isSelected(row)" [aria-label]="checkboxLabel(row)" (click)="onrowselect()" >
+      </mat-checkbox>
+    </td>
+  </ng-container>
+
+  <ng-container matColumnDef="id">
+    <th mat-header-cell *matHeaderCellDef mat-sort-header> id </th>
+    <td mat-cell *matCellDef="let element"> {{element.id}} </td>
+  </ng-container>
+
+  <ng-container matColumnDef="name">
+    <th mat-header-cell *matHeaderCellDef mat-sort-header> name </th>
+    <td mat-cell *matCellDef="let element"> {{element.name}} </td>
+  </ng-container>
+
+  <ng-container matColumnDef="quantityPerUnit">
+    <th mat-header-cell *matHeaderCellDef mat-sort-header> quantityPerUnit </th>
+    <td mat-cell *matCellDef="let element"> {{element.quantityPerUnit}} </td>
+  </ng-container>
+ 
+  <ng-container matColumnDef="unitPrice">
+    <th mat-header-cell *matHeaderCellDef mat-sort-header> unitPrice </th>
+    <td mat-cell *matCellDef="let element"> {{element.unitPrice}} </td>
+  </ng-container>
+
+  <ng-container matColumnDef="unitsInStock">
+    <th mat-header-cell *matHeaderCellDef mat-sort-header> unitsInStock </th>
+    <td mat-cell *matCellDef="let element"> {{element.unitsInStock}} </td>
+  </ng-container> 
+
+  <ng-container matColumnDef="unitsOnOrder">
+    <th mat-header-cell *matHeaderCellDef mat-sort-header> unitsOnOrder </th>
+    <td mat-cell *matCellDef="let element"> {{element.unitsOnOrder}} </td>
+  </ng-container>
+  <ng-container matColumnDef="reorderLevel">
+    <th mat-header-cell *matHeaderCellDef mat-sort-header> reorderLevel </th>
+    <td mat-cell *matCellDef="let element"> {{element.reorderLevel}} </td>
+  </ng-container>
+
+  <ng-container matColumnDef="discontinued">
+    <th mat-header-cell *matHeaderCellDef mat-sort-header> discontinued </th>
+    <td mat-cell *matCellDef="let element"> {{element.discontinued}} </td>
+  </ng-container>
+
+  <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+  <tr mat-row *matRowDef="let row; columns: displayedColumns;" (click)="selection.toggle(row);"
+    (click)="onrowselect(row)" ></tr>
+
+
+</table>
+<mat-paginator [pageSizeOptions]="[5, 10, 20]" showFirstLastButtons></mat-paginator>
