+++ src/app/products/product/product.component.scss
@@ -1,0 +1,5 @@
+ 
+table {
+    width: 100%;
+    position: relative;    
+  }
