+++ src/app/products/product/product.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ProductComponent } from './product.component';
 
-describe('ProductComponent', () => {
+describe('HopitalComponent', () => {
   let component: ProductComponent;
   let fixture: ComponentFixture<ProductComponent>;
